Integrate controller package into application package.
