+++ src/main/java/oripa/application/main/UnselectAllLinesActionListener.java
@@ -16,39 +16,43 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.controller;
+package oripa.application.main;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
 import oripa.domain.paint.PaintContextInterface;
+import oripa.domain.paint.ScreenUpdaterInterface;
 
 /**
  * @author OUCHI Koji
  *
  */
-public class SelectAllLineActionListener implements ActionListener {
+public class UnselectAllLinesActionListener implements ActionListener {
 	private final PaintContextInterface context;
+	private final ScreenUpdaterInterface screenUpdater;
 
 	/**
 	 * Constructor
 	 */
-	public SelectAllLineActionListener(final PaintContextInterface context) {
-		this.context = context;
+	public UnselectAllLinesActionListener(final PaintContextInterface aContext,
+			final ScreenUpdaterInterface updater) {
+		context = aContext;
+		screenUpdater = updater;
 	}
 
 	/*
 	 * (non Javadoc)
-	 *
+	 * 
 	 * @see
 	 * java.awt.event.ActionListener#actionPerformed(java.awt.event.ActionEvent)
 	 */
 	@Override
 	public void actionPerformed(final ActionEvent e) {
-		context.creasePatternUndo().pushUndoInfo();
-		context.getPainter().selectAllOriLines();
-		context.getCreasePattern().stream()
-				.filter(l -> l.selected).forEach(l -> context.pushLine(l));
+		context.getPainter().resetSelectedOriLines();
+		context.clear(false);
+		screenUpdater.updateScreen();
+
 	}
 
 }
