+++ src/main/java/oripa/view/main/MainFrame.java
@@ -44,15 +44,15 @@ import org.slf4j.LoggerFactory;
 import oripa.Config;
 import oripa.ORIPA;
 import oripa.application.main.DataFileAccess;
+import oripa.application.main.DeleteSelectedLinesActionListener;
 import oripa.application.main.IniFileAccess;
 import oripa.application.main.PaintContextModification;
+import oripa.application.main.SelectAllLineActionListener;
+import oripa.application.main.UnselectAllLinesActionListener;
 import oripa.appstate.StateManager;
 import oripa.bind.ButtonFactory;
 import oripa.bind.PaintActionButtonFactory;
 import oripa.bind.state.PaintBoundStateFactory;
-import oripa.controller.DeleteSelectedLinesActionListener;
-import oripa.controller.SelectAllLineActionListener;
-import oripa.controller.UnselectAllLinesActionListener;
 import oripa.doc.Doc;
 import oripa.domain.cptool.Painter;
 import oripa.domain.paint.MouseActionHolder;
